Decompose junit & module download
